+++ mcio_ctrl/envs/env_util.py
@@ -22,16 +22,16 @@ class DegreesToPixels:
     DEGREES_PER_PIXEL = 0.15
     PIXELS_PER_DEGREE = 1 / DEGREES_PER_PIXEL
 
-    def __init__(self, *, x: int = 0, y: int = 0) -> None:
+    def __init__(self, *, x: float = 0.0, y: float = 0.0) -> None:
         self.x, self.y = x, y
 
-    def update(self, *, yaw_delta: float, pitch_delta: float) -> tuple[int, int]:
+    def update(self, *, yaw_delta: float, pitch_delta: float) -> tuple[float, float]:
         """Delta arguments are in degrees. Returns the new cursor position in pixels."""
-        self.x += int(yaw_delta * self.PIXELS_PER_DEGREE)
-        self.y += int(pitch_delta * self.PIXELS_PER_DEGREE)
+        self.x += yaw_delta * self.PIXELS_PER_DEGREE
+        self.y += pitch_delta * self.PIXELS_PER_DEGREE
         return self.x, self.y
 
-    def set(self, x: int, y: int) -> None:
+    def set(self, x: float, y: float) -> None:
         """Directly set the absolute current pixel location"""
         self.x, self.y = x, y
